refactor(search): Moved extension and result view in self package
