+++ REES46/Classes/Stories/StoryViewController/SearchWidget/Extensions/SearchResultsView.swift
@@ -1,0 +1,19 @@
+import Foundation
+
+extension SearchResultsView: UITableViewDelegate, UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return results.count
+    }
+
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        let cell = tableView.dequeueReusableCell(withIdentifier: "CustomCell", for: indexPath) as! SearchResultTableViewCell
+        let result = results[indexPath.row]
+        cell.configure(with: result)
+        return cell
+    }
+
+    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+        let selectedResult = results[indexPath.row]
+        delegate?.didSelectResult(selectedResult)
+    }
+}
